refactor: remove notes store and move functions to hooks/lib
