+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Flow/notes/hooks/useClonedFlowNoteId.ts
@@ -1,0 +1,5 @@
+export const useClonedFlowNoteId = () => {
+  const getClonedFlowNoteId = () => localStorage.getItem("clonedFlowNoteId");
+
+  return { getClonedFlowNoteId };
+};
