+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Flow/notes/hooks/useCopiedFlowNote.ts
@@ -1,0 +1,13 @@
+export interface CopiedFlowNoteContent {
+  text: string;
+  color: string;
+}
+
+export const useCopiedFlowNote = () => {
+  const getCopiedFlowNote = (): CopiedFlowNoteContent | undefined => {
+    const payload = localStorage.getItem("copiedFlowNote");
+    return payload ? JSON.parse(payload) : undefined;
+  };
+
+  return { getCopiedFlowNote };
+};
